[buildsys: fix versionheader rebuild]
issue:

Updates that impact the version header require two calls to make to
actually end up in the kernel image.

reason:

The version header target is updated during the same buildsys stage as
the kernel. Hence, the kernel is not rebuild since the existing version
header is not newer. Accordingly, a second call to make is needed to
detect the updated version header and rebuild the kernel image.

fix:

Build the version header during the earlier prepare_deps buildsys
stage. That way, the version header update is detected during the main
build stage and the kernel image is update during the first make call.