[buildsys: fix lex, yacc, gperf header rules]
issue:

Build fails with a missing rule error when using lex, yacc or gperf
header files as dependencies to other targets.

reason:

The rules generated for those tools do not contain a recipe, i.e. only
specify a dependency from e.g. the .hash.cc to the .hash.h file without
defining how to build it.

This issue has been introduced with the update of the prestage handling
in 770b28dd51c4eec989b31211eb6d41a975021299. With this $(gen_rule) has
been updated, such that it strips the supplied command, only generating
a rule recipe if the result is not empty. Since the commands for the
lex, yacc and gperf header rules are empty, they don't get recipes.

fix:

Introduce the $(nop) make macro and use it as the command for the lex,
yacc and gperf rules, therefor generating rules with a recipe.